Accept the temperature with fractional part

+++ app/controllers/automation-heating.js
@@ -496,7 +496,7 @@ myAppController.controller('HeatingIdController', function($scope, $routeParams,
 						});
 
 					if(scIndex != -1) {
-						$scope.heating.input.params.roomSettings[roomId].schedule[$scope.heating.tempModal.timeline][scIndex].temp = parseInt($scope.heating.tempModal.temp.value);
+						$scope.heating.input.params.roomSettings[roomId].schedule[$scope.heating.tempModal.timeline][scIndex].temp = parseFloat($scope.heating.tempModal.temp.value);
 
 						var rows_copy = {};
 						angular.copy($scope.scheduleOptions.rows, rows_copy);
